Improve blockchain exchange parsing architecture
