+++ crates/adapters/blockchain/src/exchanges/arbitrum/sushiswap_v2.rs
@@ -34,6 +34,7 @@ pub static SUSHISWAP_V2: LazyLock<DexExtended> = LazyLock::new(|| {
         "",
         "",
         "",
+        "",
     );
     DexExtended::new(dex)
 });
