+++ crates/adapters/blockchain/src/exchanges/base/aerodrome_slipstream.rs
@@ -34,6 +34,7 @@ pub static AERODROME_SLIPSTREAM: LazyLock<DexExtended> = LazyLock::new(|| {
         "",
         "",
         "",
+        "",
     );
     DexExtended::new(dex)
 });
