+++ crates/adapters/blockchain/src/exchanges/base/baseswap_v2.rs
@@ -34,6 +34,7 @@ pub static BASESWAP_V2: LazyLock<DexExtended> = LazyLock::new(|| {
         "",
         "",
         "",
+        "",
     );
     DexExtended::new(dex)
 });
