+++ crates/adapters/blockchain/src/exchanges/base/basex.rs
@@ -34,6 +34,7 @@ pub static BASEX: LazyLock<DexExtended> = LazyLock::new(|| {
         "",
         "",
         "",
+        "",
     );
     DexExtended::new(dex)
 });
