+++ crates/adapters/blockchain/src/exchanges/base/maverick_v1.rs
@@ -34,6 +34,7 @@ pub static MAVERICK_V1: LazyLock<DexExtended> = LazyLock::new(|| {
         "",
         "",
         "",
+        "",
     );
     DexExtended::new(dex)
 });
