+++ crates/adapters/blockchain/src/exchanges/ethereum/balancer_v3.rs
@@ -34,6 +34,7 @@ pub static BALANCER_V3: LazyLock<DexExtended> = LazyLock::new(|| {
         "",
         "",
         "",
+        "",
     );
     DexExtended::new(dex)
 });
