+++ crates/adapters/blockchain/src/exchanges/ethereum/curve_finance.rs
@@ -34,6 +34,7 @@ pub static CURVE_FINANCE: LazyLock<DexExtended> = LazyLock::new(|| {
         "",
         "",
         "",
+        "",
     );
     DexExtended::new(dex)
 });
