+++ crates/adapters/blockchain/src/exchanges/ethereum/maverick_v2.rs
@@ -34,6 +34,7 @@ pub static MAVERICK_V2: LazyLock<DexExtended> = LazyLock::new(|| {
         "",
         "",
         "",
+        "",
     );
     DexExtended::new(dex)
 });
