+++ crates/adapters/blockchain/src/exchanges/ethereum/pancakeswap_v3.rs
@@ -34,6 +34,7 @@ pub static PANCAKESWAP_V3: LazyLock<DexExtended> = LazyLock::new(|| {
         "",
         "",
         "",
+        "",
     );
     DexExtended::new(dex)
 });
