+++ crates/adapters/blockchain/src/exchanges/mod.rs
@@ -13,7 +13,7 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-use nautilus_model::defi::{Blockchain, DexType};
+use nautilus_model::defi::{Blockchain, Chain, DexType};
 
 use crate::exchanges::{
     arbitrum::ARBITRUM_DEX_EXTENDED_MAP, base::BASE_DEX_EXTENDED_MAP,
@@ -24,6 +24,7 @@ pub mod arbitrum;
 pub mod base;
 pub mod ethereum;
 pub mod extended;
+mod parsing;
 
 /// Returns a map of all DEX names to Dex instances across all chains
 #[must_use]
@@ -53,4 +54,23 @@ pub fn get_supported_dexes_for_chain(blockchain: Blockchain) -> Vec<String> {
         .into_iter()
         .map(|dex_type| format!("{dex_type}"))
         .collect()
+}
+
+/// Attempts to match a DEX name in a case-insensitive manner.
+pub fn find_dex_type_case_insensitive(dex_name: &str, chain: &Chain) -> Option<DexType> {
+    let supported_dexes = get_supported_dexes_for_chain(chain.name);
+
+    // First try exact match (for performance)
+    if let Some(dex_type) = DexType::from_dex_name(dex_name) {
+        return Some(dex_type);
+    }
+
+    // Try case-insensitive match
+    for supported_dex in supported_dexes {
+        if supported_dex.to_lowercase() == dex_name.to_lowercase() {
+            return DexType::from_dex_name(&supported_dex);
+        }
+    }
+
+    None
 }
