+++ crates/adapters/blockchain/src/exchanges/parsing/uniswap_v3/mod.rs
@@ -13,28 +13,10 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-use std::sync::LazyLock;
-
-use nautilus_model::defi::{
-    chain::chains,
-    dex::{AmmType, Dex, DexType},
-};
-
-use crate::exchanges::extended::DexExtended;
-
-/// Curve Finance DEX on Arbitrum.
-pub static CURVE_FINANCE: LazyLock<DexExtended> = LazyLock::new(|| {
-    let dex = Dex::new(
-        chains::ARBITRUM.clone(),
-        DexType::CurveFinance,
-        "0xb17b674D9c5CB2e441F8e196a2f048A81355d031",
-        1413161,
-        AmmType::StableSwap,
-        "",
-        "",
-        "",
-        "",
-        "",
-    );
-    DexExtended::new(dex)
-});
+pub mod burn;
+pub mod collect;
+pub mod initialize;
+pub mod mint;
+pub mod pool_created;
+pub mod swap;
+pub mod trade_data;
